requirements.txt
alabaster==0.7.12
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
docutils==0.16
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==5.6.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pluggy==0.13.1
py==1.9.0
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1

# TODO Clean it up